+++ model/recommend.go
@@ -1,15 +1,15 @@
 package model
 
 type Evaluation interface {
-	Score(Solution) float32
+	Score(Solution) float64
 }
 
 type EvaluationWithWeight struct {
 	EvaluationWay Evaluation
-	Weight        float32
+	Weight        float64
 }
 
-func MakeEvaluationWithWeight(eval Evaluation, weight float32) EvaluationWithWeight {
+func MakeEvaluationWithWeight(eval Evaluation, weight float64) EvaluationWithWeight {
 	return EvaluationWithWeight{eval, weight}
 }
 
@@ -27,11 +27,11 @@ func (s Sense) AddEWW(eww EvaluationWithWeight) {
 	s.Evals = append(s.Evals, eww)
 }
 
-func (s Sense) Add(eval Evaluation, weight float32) {
+func (s Sense) Add(eval Evaluation, weight float64) {
 	s.AddEWW(MakeEvaluationWithWeight(eval, weight))
 }
 
-func (s Sense) Score(sol Solution) (score float32) {
+func (s Sense) Score(sol Solution) (score float64) {
 	for _, eval := range s.Evals {
 		score += eval.EvaluationWay.Score(sol) * eval.Weight
 	}
